Split latest_reports() into an internal and external version.
