+++ components/server/src/external/routes/measurement.py
@@ -9,7 +9,6 @@ import bottle
 from pymongo.database import Database
 
 from database.measurements import insert_new_measurement, latest_measurement
-from database.reports import latest_reports
 from server_utilities.type import MetricId, SourceId
 from shared.model.measurement import Measurement
 from shared.model.metric import Metric
@@ -17,6 +16,7 @@ from shared.model.metric import Metric
 from ..database import sessions
 from ..database.datamodels import latest_datamodel
 from ..database.measurements import count_measurements, measurements_by_metric
+from ..database.reports import latest_reports
 from ..model.data import SourceData
 from ..utils.functions import report_date_time
