+++ components/server/src/external/routes/subject.py
@@ -4,13 +4,11 @@ from datetime import datetime, timedelta, timezone
 import bottle
 from pymongo.database import Database
 
-from database.reports import latest_reports
 from server_utilities.type import MetricId, ReportId, SubjectId
 
 from ..database.datamodels import default_subject_attributes, latest_datamodel
 from ..database.measurements import measurements_by_metric
-from ..database.reports import insert_new_report
-from ..database.reports import metrics_of_subject
+from ..database.reports import insert_new_report, metrics_of_subject, latest_reports
 from ..model.actions import copy_subject, move_item
 from ..model.data import ReportData, SubjectData
 from ..utils.functions import report_date_time, sanitize_html, uuid
