+++ components/server/src/internal/database/reports.py
@@ -2,11 +2,18 @@
 
 from pymongo.database import Database
 
+from database.filters import DOES_NOT_EXIST
 from server_utilities.type import MetricId
 from shared.model.metric import Metric
-from database.reports import latest_reports
+from shared.model.report import Report
 
 from .datamodels import latest_datamodel
+
+
+def latest_reports(database: Database, data_model: dict) -> list[Report]:
+    """Return the latest, undeleted, reports in the reports collection."""
+    report_dicts = database.reports.find({"last": True, "deleted": DOES_NOT_EXIST})
+    return [Report(data_model, report_dict) for report_dict in report_dicts]
 
 
 def latest_metric(database: Database, metric_uuid: MetricId) -> Metric | None:
